feat: part 2 is consistent with structure
